Add task test cases files api to list files and refactoring to the code
