+++ inginious/frontend/pages/course_admin/task_edit.py
@@ -52,12 +52,11 @@ class CourseEditTask(INGIniousAdminPage):
         available_filetypes = self.task_factory.get_available_task_file_extensions()
 
         additional_tabs = self.plugin_manager.call_hook('task_editor_tab', course=course, taskid=taskid,
-                                                        task_data = task_data, template_helper = self.template_helper)
+                                                        task_data=task_data, template_helper=self.template_helper)
 
         additional_footer = self.plugin_manager.call_hook('task_editor_footer', course=course, taskid=taskid,
                                                           task_data=task_data, template_helper=self.template_helper)
 
-        print(additional_tabs)
 
         return self.template_helper.get_renderer().course_admin.task_edit(
             course,
