+++ inginious/frontend/plugins/grader_generator/pages/grader_templates.html
